+++ index.html
@@ -226,7 +226,6 @@
       </div>
     </section>
   </template>
-  <script scr="js/main.js"></script>
-  <script src="js/functions.js"></script>
+  <script src="js/main.js" type="module"></script>
 </body>
 </html>
